Move type adaptation tests into a separate file and add a unified test for all supported literal type adaptation.
